Add Arbitrary distribution (#1704)
* rename from_file to arbitrary

* add arbitrary distribution

* use logit if bounds supplied

* fix bugs

* fix bug in isin checking

* fix name space clashes, whitespace

* fix doc string
