+++ pycbc/distributions/__init__.py
@@ -18,7 +18,7 @@ probability density function of distributions.
 """
 
 from pycbc.distributions.angular import *
-from pycbc.distributions.from_file import *
+from pycbc.distributions.arbitrary import *
 from pycbc.distributions.gaussian import *
 from pycbc.distributions.power_law import *
 from pycbc.distributions.sky_location import *
@@ -26,6 +26,7 @@ from pycbc.distributions.uniform import *
 
 # a dict of all available distributions
 distribs = {
+    Arbitrary.name : Arbitrary,
     FromFile.name : FromFile,
     Gaussian.name : Gaussian,
     UniformPowerLaw.name : UniformPowerLaw,
